ms-authentification/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.61.1
PyJWT==2.8.0
six==1.16.0

#pip install -r requirements.txt
